Move record and delete-records to top-level commands
- Create new fulcra_api/cli/record.py module
- Move record command from data-type subgroup to top-level
- Move delete-records command from data-type subgroup to top-level
- Update command invocations in help text (fulcra record vs fulcra data-type record)
- Simpler CLI structure: fulcra record, fulcra delete-records
