+++ fulcra_api/cli/__init__.py
@@ -25,6 +25,7 @@ from .commands import (
 )
 from .data_types import data_type
 from .files import file
+from .record import delete_records, record
 from .share import share
 from .tags import tag
 from .utils import ensure_config_directory, load_creds, save_creds, pass_fulcra_api
@@ -52,6 +53,8 @@ cli.add_command(tag)
 cli.add_command(data_type)
 cli.add_command(file)
 cli.add_command(share)
+cli.add_command(record)
+cli.add_command(delete_records)
 
 cli.add_command(list_calendars)
 cli.add_command(list_calendar_events)
